Move error declaration in a seperate file.
